requirements.txt
paho-mqtt
dash
python-dotenv
pandas
